.tmp/langfuse-bootstrap/requirements.txt
langfuse
openai
python-dotenv
